+++ engine/MainClass.h
@@ -21,7 +21,7 @@ namespace Engine {
     class ShaderCompiler;
 
     class ComplexRenderGraphBuilder;
-    class RenderGraph2;
+    class RenderGraph;
     enum class RGTextureHandle : int32_t;
 
     class MainClass {
@@ -61,7 +61,7 @@ namespace Engine {
         std::shared_ptr<Input> GetInputSystem() const;
         std::shared_ptr<ShaderCompiler> GetShaderCompiler() const;
 
-        void SetRenderGraph(std::unique_ptr<RenderGraph2> render_graph, RGTextureHandle final_color_attachment_id);
+        void SetRenderGraph(std::unique_ptr<RenderGraph> render_graph, RGTextureHandle final_color_attachment_id);
 
     protected:
         // XXX: window must destroyed before renderer. Because the window has some AllocatedImage2D. So the permutation
@@ -76,7 +76,7 @@ namespace Engine {
         std::shared_ptr<Input> input{};
         std::shared_ptr<ShaderCompiler> shader_compiler{};
 
-        std::unique_ptr<RenderGraph2> render_graph{};
+        std::unique_ptr<RenderGraph> render_graph{};
         RGTextureHandle m_final_color_attachment_id{0};
 
         static std::weak_ptr<MainClass> m_instance;
