+++ engine/Render/Pipeline/RenderGraph/RenderGraphBuilder.h
@@ -1,167 +1,111 @@
-#ifndef PIPELINE_RENDERGRAPH_RENDERGRAPHBUILDER_INCLUDED
-#define PIPELINE_RENDERGRAPH_RENDERGRAPHBUILDER_INCLUDED
+#ifndef PIPELINE_RENDERGRAPH2_RENDERGRAPHBUILDER2
+#define PIPELINE_RENDERGRAPH2_RENDERGRAPHBUILDER2
 
-#include <functional>
-#include <memory>
-
-#include "Render/AttachmentUtils.h"
 #include "Render/Memory/MemoryAccessTypes.h"
 #include "Render/Memory/RenderTargetTexture.h"
 
-#include "RGAttachmentDesc.h"
-
 namespace Engine {
-    class RenderSystem;
-    class GUISystem;
-    class DeviceBuffer;
     class RenderGraph;
-    class GraphicsCommandBuffer;
-    class TransferCommandBuffer;
-    class ComputeCommandBuffer;
+    class RenderGraphPass;
+    class RenderSystem;
+    class DeviceBuffer;
+    class RRTTHandle;
 
-    /**
-     * @brief Helper class for building a `RenderGraph`.
-     *
-     * Helps to resolve data dependencies between passes and set up barriers.
-     */
     class RenderGraphBuilder {
         struct impl;
+        RenderSystem &system;
         std::unique_ptr<impl> pimpl;
-
-    protected:
-        RenderSystem &m_system;
 
     public:
         RenderGraphBuilder(RenderSystem &system);
         ~RenderGraphBuilder();
 
         /**
-         * @brief Register a new image texture to manage its access by the internal memo system.
+         * @brief Register a new image texture to manage its access internally.
          *
          * The lifetime of this texture is managed by the caller.
-         * This method facilitates persistent data usage between frames (for TAA, for example).
+         * This method facilitates persistent data usage between frames
+         * (for TAA, for example).
          *
-         * @return a handle to the managed resource, used in the render graph internally.
-         * External resources will have negative handles.
+         * The reference is non const as new views might be generated during
+         * rendering.
+         *
+         * @return a handle to the managed resource, used in the render graph
+         * internally. External resources will have negative handles.
          */
-        int32_t ImportExternalResource(
-            const RenderTargetTexture &texture, MemoryAccessTypeImageBits prev_access = MemoryAccessTypeImageBits::None
+        [[nodiscard]]
+        RGTextureHandle ImportExternalResource(
+            RenderTargetTexture &texture, MemoryAccessTypeImageBits prev_access = MemoryAccessTypeImageBits::None
+        );
+        [[nodiscard]]
+        RGTextureHandle ImportExternalResource(
+            RRTTHandle texture, MemoryAccessTypeImageBits prev_access = MemoryAccessTypeImageBits::None
         );
 
         /**
-         * @brief Register a new buffer to manage its access by the internal memo system.
+         * @brief Register a new buffer to manage its access internally.
          *
-         * @return a handle to the managed resource, used in the render graph internally.
-         * External resources will have negative handles.
+         * @return a handle to the managed resource, used in the render graph
+         * internally. External resources will have negative handles.
          */
-        int32_t ImportExternalResource(
+        [[nodiscard]]
+        RGBufferHandle ImportExternalResource(
             const DeviceBuffer &buffer, MemoryAccessTypeBuffer prev_access = {MemoryAccessTypeBufferBits::None}
         );
 
         /**
-         * @brief Request a new render target texture to be created when compiling the
-         * render graph.
+         * @brief Request a new render target texture to be created when
+         * compiling the render graph.
          *
-         * Such resources will have their lifetime managed automatically by the compiled
-         * render graph.
+         * Such resources will have their lifetime managed automatically by the
+         * compiled render graph.
+         * Render target allocated by this method is considered transient. They
+         * cannot be used across frames. If you wish to use a persistent render
+         * texture, import it with `ImportExternalResource` method.
          *
-         * @return a handle to the managed resource, used in the render graph internally.
-         * Internal resouces will have positive handles.
+         * @return a handle to the managed resource, used in the render graph
+         * internally. Internal resouces will have positive handles.
          */
-        int32_t RequestRenderTargetTexture(
+        [[nodiscard]]
+        RGTextureHandle RequestRenderTargetTexture(
             RenderTargetTexture::RenderTargetTextureDesc texture_description,
-            RenderTargetTexture::SamplerDesc sampler_description
+            RenderTargetTexture::SamplerDesc sampler_description,
+            std::string_view name = ""
         ) noexcept;
 
         /**
-         * @brief Mark an image to be used in the following pass.
+         * @brief Request a new render target texture to be created when
+         * compiling the render graph.
          *
-         * @details actual access to the image is not determined until the next pass
-         * is recorded. You can only have one bit of access, as the access is tied
-         * to the current image layout.
-         */
-        void UseImage(int32_t texture_handle, MemoryAccessTypeImageBits access);
-
-        /**
-         * @brief Mark a buffer to be used in the following pass.
-         */
-        void UseBuffer(int32_t buffer_handle, MemoryAccessTypeBuffer access);
-
-        /**
-         * @brief Record a pass which includes draw calls.
+         * This render target texture will be resizable and managed by the
+         * manager attached to the current render system. However, its
+         * ownership is mantained by the built render graph, and will be release
+         * when destructing.
          *
-         * To use this method, you have to manually set up render targets within the `pass` function.
+         * @return a handle to the managed resource, used in the render graph
+         * internally. Internal resouces will have positive handles.
          */
-        void RecordRasterizerPassWithoutRT(std::function<void(GraphicsCommandBuffer &, const RenderGraph &)> pass);
-
-        /**
-         * @brief Record a pass which includes draw calls.
-         *
-         * Depth attachment is defaulted to null.
-         * This method automatically begins and terminates rendering.
-         * Its extent is set to swapchain extent.
-         */
-        void RecordRasterizerPass(
-            RGAttachmentDesc color,
-            std::function<void(GraphicsCommandBuffer &, const RenderGraph &)> pass,
-            const std::string &name = ""
+        [[nodiscard]]
+        RGTextureHandle RequestResizableRenderTargetTexture(
+            RenderTargetTexture::RenderTargetTextureDesc texture_description,
+            RenderTargetTexture::SamplerDesc sampler_description,
+            float scale_x = 1.0f,
+            float scale_y = 1.0f,
+            std::string_view name = ""
         );
 
         /**
-         * @brief Record a pass which includes draw calls.
+         * @brief Add a pass to this render graph.
          *
-         * This method automatically begins and terminates rendering.
-         * Its extent is set to swapchain extent.
+         * Use `RenderGraphPassBuilder` to construct the pass.
          */
-        void RecordRasterizerPass(
-            RGAttachmentDesc color,
-            RGAttachmentDesc depth,
-            std::function<void(GraphicsCommandBuffer &, const RenderGraph &)> pass,
-            const std::string &name = ""
-        );
+        void AddPass(RenderGraphPass &&pass) noexcept;
 
         /**
-         * @brief Record a pass which includes draw calls.
-         *
-         * This method automatically begins and terminates rendering.
-         * Its extent is set to swapchain extent.
-         */
-        void RecordRasterizerPass(
-            std::initializer_list<RGAttachmentDesc> colors,
-            RGAttachmentDesc depth,
-            std::function<void(GraphicsCommandBuffer &, const RenderGraph &)> pass,
-            const std::string &name = ""
-        );
-
-        /**
-         * @brief Record a pass with transfer commands.
-         */
-        void RecordTransferPass(
-            std::function<void(TransferCommandBuffer &, const RenderGraph &)> pass, const std::string &name = ""
-        );
-
-        /**
-         * @brief Record a pass with compute shader dispatches.
-         */
-        void RecordComputePass(
-            std::function<void(ComputeCommandBuffer &, const RenderGraph &)> pass, const std::string &name = ""
-        );
-
-        /**
-         * @brief Create a RenderGraph and reset internal states to default.
+         * @brief Construct a render graph according to the passes.
          */
         std::unique_ptr<RenderGraph> BuildRenderGraph();
-
-        /**
-         * @brief Build a default render graph.
-         *
-         * The render graph will have defaulted attachments.
-         * Color and depth attachments will always have handle value 0 and 1 respectively.
-         */
-        std::unique_ptr<RenderGraph> BuildDefaultRenderGraph(
-            uint32_t width, uint32_t height, GUISystem *gui_system = nullptr
-        );
     };
 } // namespace Engine
 
-#endif // PIPELINE_RENDERGRAPH_RENDERGRAPHBUILDER_INCLUDED
+#endif // PIPELINE_RENDERGRAPH2_RENDERGRAPHBUILDER2
