+++ test/skybox_test.cpp
@@ -148,7 +148,7 @@ int main(int argc, char **argv) {
     }
 
     // Dummy texture for presenting
-    RenderGraphBuilder2 rgb{*rsys};
+    RenderGraphBuilder rgb{*rsys};
     auto crt = rgb.RequestRenderTargetTexture(
         RenderTargetTexture::RenderTargetTextureDesc{
             .dimensions = 2,
@@ -187,7 +187,7 @@ int main(int argc, char **argv) {
                  AttachmentUtils::StoreOperation::DontCare,
                  AttachmentUtils::DepthClearValue{1.0f, 0U}}
             )
-            .SetRasterizerPassFunction([rsys, camera](GraphicsCommandBuffer &cb, const RenderGraph2 &) -> void {
+            .SetRasterizerPassFunction([rsys, camera](GraphicsCommandBuffer &cb, const RenderGraph &) -> void {
                 glm::mat3 view_matrix = glm::mat3(camera->GetViewMatrix());
                 glm::mat4 pv = camera->GetProjectionMatrix() * glm::mat4(view_matrix);
                 rsys->GetSceneDataManager().DrawSkybox(
